Add solver for nonograms.com
A simple solver for the puzzles on the https://www.nonograms.org/ site.
The solver currently only works on small to medium sized black & white
puzzles. It is currently too slow for for complex large puzzles and it
doesn't handle color puzzles.

Implemented some ad-hoc optimizations. However, they were not very
effective.

The solver takes screenshots of the display using the mss library and
interacts with the puzzle using the PyAutoGUI library.